[PAGOPA-2696] chore: refactoring clients package
